llm_ptq: dedup preset-discovery helper into example_utils
hf_ptq.py and multinode_ptq.py both live in examples/llm_ptq/ and both imported
from example_utils already, so move the shared _PresetCfgChoices helper (plus the
preset-dir constants, _QFORMAT_ALIASES table, _KV_NONE sentinel, the constructed
QUANT_CFG_CHOICES / KV_QUANT_CFG_CHOICES mappings and the sentinel-collision
guard) into example_utils.py and import them in both scripts. The hf_ptq-only
auto_quantize helpers (_AUTO_QUANTIZE_QFORMATS, _canonical_qformat) and
_kv_cfg_uses_constant_amax stay local. No behavior change.

The megatron_bridge example keeps its own copy: it is a separate example
directory and does not share example_utils.
