+++ examples/llm_ptq/hf_ptq.py
@@ -18,7 +18,6 @@ import copy
 import random
 import time
 import warnings
-from collections.abc import Iterator, Mapping
 from pathlib import Path
 from typing import Any
 
@@ -28,6 +27,10 @@ from accelerate.hooks import remove_hook_from_module
 from cast_mxfp4_to_nvfp4 import apply_to_model as apply_cast_mxfp4_to_nvfp4
 from cast_mxfp4_to_nvfp4 import force_weight_quantizers_static
 from example_utils import (
+    _KV_NONE,
+    _QFORMAT_ALIASES,
+    KV_QUANT_CFG_CHOICES,
+    QUANT_CFG_CHOICES,
     build_quant_cfg,
     copy_custom_model_files,
     create_vlm_calibration_loop,
@@ -56,7 +59,7 @@ from transformers import (
 import modelopt.torch.opt as mto
 import modelopt.torch.quantization as mtq
 import modelopt.torch.sparsity as mts
-from modelopt.recipe import ModelOptPTQRecipe, load_config, load_recipe
+from modelopt.recipe import ModelOptPTQRecipe, load_recipe
 from modelopt.torch.export import (
     export_hf_checkpoint,
     export_hf_vllm_fq_checkpoint,
@@ -67,12 +70,7 @@ from modelopt.torch.export import (
     save_expert_token_count_table,
 )
 from modelopt.torch.export.model_utils import get_language_model_from_vl, is_multimodal_model
-from modelopt.torch.opt.config_loader import BUILTIN_CONFIG_ROOT
-from modelopt.torch.quantization.config import (
-    QuantizeConfig,
-    _default_disabled_quantizer_cfg,
-    need_calibration,
-)
+from modelopt.torch.quantization.config import _default_disabled_quantizer_cfg, need_calibration
 from modelopt.torch.quantization.plugins.accelerate import init_quantized_weights
 from modelopt.torch.quantization.utils import is_quantized
 from modelopt.torch.speculative.eagle.utils import (
@@ -92,44 +90,6 @@ from modelopt.torch.utils.vlm_dataset_utils import get_vlm_dataset_dataloader
 RAND_SEED = 1234
 
 
-# Preset directories under modelopt_recipes/ that back the --qformat and
-# --kv_cache_qformat CLI vocabularies. Each ``*.yaml`` file in these directories is
-# automatically discovered and exposed as a valid CLI value via _PresetCfgChoices,
-# so no code change in this script is required when a YAML is added or removed.
-# This is deliberate: every preset YAML is CLI-exposed, there is no separate
-# allow-list — the directory listing is the policy.
-#
-# That said, prefer NOT to add new YAMLs to these preset directories either. The
-# long-term direction is to retire --qformat / --kv_cache_qformat entirely in favour
-# of --recipe, which accepts a full PTQ recipe (see modelopt_recipes/general/ptq/
-# and modelopt/recipe/). New quantization configurations should be authored as
-# recipes, not as preset entries.
-_QFORMAT_PRESET_DIR = "configs/ptq/presets/model"
-_KV_QFORMAT_PRESET_DIR = "configs/ptq/presets/kv"
-
-# Backward-compat short names → canonical preset basename. These aliases predate the
-# YAML-driven discovery below and remain accepted so existing scripts keep working.
-#
-# DO NOT add new entries here. New quantization formats must be exposed via their YAML
-# basename under modelopt_recipes/configs/ptq/presets/model/ — the directory listing is
-# the canonical CLI vocabulary. This table exists solely to keep pre-existing short
-# names (and the scripts/docs that hardcode them) working through deprecation, and
-# should only ever shrink.
-_QFORMAT_ALIASES: dict[str, str] = {
-    "int8_sq": "int8_smoothquant",
-    "int8_wo": "int8_weight_only",
-    "w4a8_awq": "w4a8_awq_beta",
-    "nvfp4_awq": "nvfp4_awq_lite",
-    "nvfp4_mse": "nvfp4_w4a4_weight_mse_fp8_sweep",
-    "nvfp4_local_hessian": "nvfp4_w4a4_weight_local_hessian",
-    "fp8_pb_wo": "fp8_2d_blockwise_weight_only",
-    "fp8_pc_pt": "fp8_per_channel_per_token",
-}
-
-# Sentinel value for ``--kv_cache_qformat`` meaning "no KV cache quantization".
-_KV_NONE = "none"
-
-
 def _kv_cfg_uses_constant_amax(kv_quant_cfg: list[dict[str, Any]]) -> bool:
     """Return True if this KV cfg pins ``use_constant_amax`` on the bmm quantizer.
 
@@ -147,72 +107,6 @@ def _kv_cfg_uses_constant_amax(kv_quant_cfg: list[dict[str, Any]]) -> bool:
         return bool(cfg.get("use_constant_amax"))
     return False
 
-
-class _PresetCfgChoices(Mapping[str, dict[str, Any]]):
-    """Lazy mapping of qformat names → quant_cfg dicts loaded from preset YAMLs.
-
-    Iterates the YAML files in ``modelopt_recipes/<subdir>/`` to populate the set
-    of available qformat names; the supplied ``aliases`` table maps additional
-    short names onto canonical preset basenames. Loading happens on first access
-    and is memoised so repeated lookups are cheap.
-    """
-
-    def __init__(self, subdir: str, aliases: Mapping[str, str] | None = None):
-        self._subdir = subdir
-        self._aliases: dict[str, str] = dict(aliases or {})
-        self._presets: set[str] = set()
-        for entry in BUILTIN_CONFIG_ROOT.joinpath(subdir).iterdir():
-            name = entry.name
-            if name.endswith((".yaml", ".yml")):
-                self._presets.add(name.rsplit(".", 1)[0])
-        # Aliases that point at non-existent presets would silently fail at access
-        # time; surface this at import instead.
-        for alias, target in self._aliases.items():
-            if target not in self._presets:
-                raise ValueError(
-                    f"Alias {alias!r} points at preset {target!r} which is not present "
-                    f"under modelopt_recipes/{subdir}/."
-                )
-        self._cache: dict[str, dict[str, Any]] = {}
-
-    def _canonical(self, key: str) -> str | None:
-        if key in self._presets:
-            return key
-        return self._aliases.get(key)
-
-    def __contains__(self, key: object) -> bool:
-        return isinstance(key, str) and self._canonical(key) is not None
-
-    def __getitem__(self, key: str) -> dict[str, Any]:
-        canon = self._canonical(key)
-        if canon is None:
-            raise KeyError(key)
-        if canon not in self._cache:
-            self._cache[canon] = load_config(
-                f"{self._subdir}/{canon}", schema_type=QuantizeConfig
-            ).model_dump(exclude_unset=True)
-        # Deepcopy on retrieval so callers can freely mutate the returned config
-        # (append per-model overrides, etc.) without poisoning the cached entry.
-        return copy.deepcopy(self._cache[canon])
-
-    def __iter__(self) -> Iterator[str]:
-        yield from sorted(self._presets | set(self._aliases))
-
-    def __len__(self) -> int:
-        return len(self._presets) + len(self._aliases)
-
-
-QUANT_CFG_CHOICES: Mapping[str, dict[str, Any]] = _PresetCfgChoices(
-    _QFORMAT_PRESET_DIR, _QFORMAT_ALIASES
-)
-KV_QUANT_CFG_CHOICES: Mapping[str, dict[str, Any]] = _PresetCfgChoices(_KV_QFORMAT_PRESET_DIR)
-
-# Guard against a future ``none.yaml`` (or alias) colliding with the disable sentinel:
-# argparse would silently allow both, but the runtime branch on ``!= _KV_NONE`` would
-# become ambiguous and the user couldn't reach the real preset.
-assert _KV_NONE not in KV_QUANT_CFG_CHOICES, (
-    f"_KV_NONE sentinel {_KV_NONE!r} collides with a KV preset; rename the preset."
-)
 
 # Formats supported by mtq.auto_quantize unified-checkpoint export.
 #
